improvements to GCsummary
