+++ ui/src/main/java/ucar/nc2/ui/grib/GribIndexPanel.java
@@ -312,7 +312,7 @@ public class GribIndexPanel extends JPanel {
 
     private void show(Grib2Record gr2, Formatter f) {
       try {
-        Grib2CollectionPanel.showCompleteGribRecord(f, indexFile, gr2, cust2);
+        Grib2Show.showCompleteGribRecord(f, indexFile, gr2, cust2);
       } catch (IOException e) {
         e.printStackTrace();
       }
